Flextable format for pivot_table

Flextable format for pivot_table

pivot_format(
  pivot,
  background = "#81A1C1",
  color = "#FFFFFF",
  border = "#FFFFFF",
  font_size = 14,
  font_name = NULL,
  labels = pivot_labels(),
  formatter = pivot_formatter(),
  zebra_style = c("classic", "stats", "none"),
  zebra_color = "#ECEFF4",
  drop_stats = FALSE,
  keep_data = TRUE
)

Arguments

pivot
A pivot_table object.

background
Background color for the header and column(s) containing row(s).

color
Text color for the header and column(s) containing row(s).

border
Border color (applies to all table).

font_size
Font size (applies to all table).

font_name
Font name (applies to all table).

labels
Custom labels for statistics, see pivot_labels.

formatter
Function to format content, see pivot_formatter.

zebra_style
Add zebra theme to table.

zebra_color
Color to use for zebra theme.

drop_stats
Drop the stats column, can be useful if have only one stat to show.

keep_data
Keep data as attribute, this can be useful to retrieve the data from which the table was formatted.

Value

a flextable object.

Examples

library(flexpivot)
library(magrittr)
data("nobel_laureates")

# With two variables
nobel_laureates %>%
  pivot_table("category", "gender") %>%
  pivot_format()

#> a flextable object.
#> col_keys: `category`, `Statistic`, `male`, `female`, `<missing>`, `Total` 
#> header has 2 row(s) 
#> body has 28 row(s) 
#> original dataset sample: 
#>            category Statistic  male female <missing> Total
#> 1         Chemistry         N   179      5         0   184
#> 2         Chemistry         % 18.8%   0.5%        0% 19.4%
#> 3         Chemistry     Col % 20.6%   9.3%        0% 19.4%
#> 4         Chemistry     Row % 97.3%   2.7%        0%  100%
#> 5 Economic Sciences         N    82      2         0    84

# Only count
nobel_laureates %>%
  pivot_table("category", "gender", stats = "n") %>%
  pivot_format(drop_stats = TRUE)

#> a flextable object.
#> col_keys: `category`, `male`, `female`, `<missing>`, `Total` 
#> header has 2 row(s) 
#> body has 7 row(s) 
#> original dataset sample: 
#>            category Statistic male female <missing> Total
#> 1         Chemistry         N  179      5         0   184
#> 2 Economic Sciences         N   82      2         0    84
#> 3        Literature         N  101     15         0   116
#> 4             Peace         N   90     17        27   134
#> 5           Physics         N  210      3         0   213

# Only percent
nobel_laureates %>%
  pivot_table("category", "gender", stats = "p") %>%
  pivot_format()

#> a flextable object.
#> col_keys: `category`, `Statistic`, `male`, `female`, `<missing>`, `Total` 
#> header has 2 row(s) 
#> body has 7 row(s) 
#> original dataset sample: 
#>            category Statistic  male female <missing> Total
#> 1         Chemistry         % 18.8%   0.5%        0% 19.4%
#> 2 Economic Sciences         %  8.6%   0.2%        0%  8.8%
#> 3        Literature         % 10.6%   1.6%        0% 12.2%
#> 4             Peace         %  9.5%   1.8%      2.8% 14.1%
#> 5           Physics         % 22.1%   0.3%        0% 22.4%

# Two variable as rows
nobel_laureates %>%
  pivot_table(c("birth_continent", "gender"), "category", stats = "p") %>%
  pivot_format()

#> a flextable object.
#> col_keys: `birth_continent`, `gender`, `Statistic`, `Chemistry`, `Economic Sciences`, `Literature`, `Peace`, `Physics`, `Physiology or Medicine`, `Total` 
#> header has 2 row(s) 
#> body has 32 row(s) 
#> original dataset sample: 
#>   birth_continent    gender Statistic Chemistry Economic Sciences Literature
#> 1          Europe      male         %     10.6%                3%       7.5%
#> 2          Europe    female         %      0.2%              0.1%       0.9%
#> 3          Europe <missing>         %        0%                0%         0%
#> 4          Europe     Total         %     10.8%              3.2%       8.4%
#> 5         Oceania      male         %      0.3%                0%         0%
#>   Peace Physics Physiology or Medicine Total
#> 1    5%   11.8%                  11.5% 49.5%
#> 2  0.5%    0.2%                   0.5%  2.5%
#> 3    0%      0%                     0%    0%
#> 4  5.6%     12%                    12%   52%
#> 5  0.2%    0.2%                   0.7%  1.5%

# Without cols
pivot_format(pivot_table(nobel_laureates, rows = "category"))

#> a flextable object.
#> col_keys: `category`, `N`, `%` 
#> header has 1 row(s) 
#> body has 7 row(s) 
#> original dataset sample: 
#>            category   N     %
#> 1         Chemistry 184 19.4%
#> 2 Economic Sciences  84  8.8%
#> 3        Literature 116 12.2%
#> 4             Peace 134 14.1%
#> 5           Physics 213 22.4%

pivot_format(pivot_table(nobel_laureates, rows = c("category", "gender")))

#> a flextable object.
#> col_keys: `category`, `gender`, `N`, `%` 
#> header has 1 row(s) 
#> body has 28 row(s) 
#> original dataset sample: 
#>            category    gender   N     %
#> 1         Chemistry      male 179 18.8%
#> 2         Chemistry    female   5  0.5%
#> 3         Chemistry <missing>   0    0%
#> 4         Chemistry     Total 184 19.4%
#> 5 Economic Sciences      male  82  8.6%

# Customize
nobel_laureates %>%
  pivot_table("category") %>%
  pivot_format(
    background = "#A3BE8C",
    labels = pivot_labels(n = "Count", p = "Percentage")
  )

#> a flextable object.
#> col_keys: `category`, `Count`, `Percentage` 
#> header has 1 row(s) 
#> body has 7 row(s) 
#> original dataset sample: 
#>            category Count Percentage
#> 1         Chemistry   184      19.4%
#> 2 Economic Sciences    84       8.8%
#> 3        Literature   116      12.2%
#> 4             Peace   134      14.1%
#> 5           Physics   213      22.4%